pyproject.toml
[project]
name = "assigment4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "fastapi>=0.134.0",
    "httpx>=0.28.1",
    "openai>=1.58.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "redis>=7.2.1",
    "reportlab>=4.2.5",
    "scalar-fastapi>=1.6.2",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
